Client E command, server ending switch
Added the if statement to run if the client is to send an E command.
Removed a closing bracket on line 143 of the server and added one on
line 114 to close the switch. I think I got that right, but it seemed
that the switch wasn't being closed and so the break command was voiding
all of the packet code.