Dropped Packets counting and report
Added NumOfPackets to handle sequence number of 'A' packets that aren't
received. Also made function printResults, that currently just prints
the data related to dropped packets.

Currently, the last packets are frequently getting dropped by the
client, probably because it is finishing the loop before receiving. I
tried doing another check at the end of it all, which helped but was
imperfect. That change is not in this commit.